# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/enrollments": {
      "get": {
        "operationId": "EnrollmentsController_findAll",
        "parameters": [
          {
            "name": "studentId",
            "required": false,
            "in": "query",
            "description": "Filter results by student identifier",
            "schema": {
              "minimum": 1,
              "example": 501,
              "type": "number"
            }
          },
          {
            "name": "classGroupId",
            "required": false,
            "in": "query",
            "description": "Filter results by class group identifier",
            "schema": {
              "minimum": 1,
              "example": 32,
              "type": "number"
            }
          },
          {
            "name": "schoolYearId",
            "required": false,
            "in": "query",
            "description": "Filter results by school year identifier",
            "schema": {
              "minimum": 1,
              "example": 2026,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size for pagination",
            "schema": {
              "minimum": 1,
              "example": 25,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Enrollments"
        ]
      },
      "post": {
        "operationId": "EnrollmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnrollmentDto"
              },
              "examples": {
                "default": {
                  "summary": "Enroll student into class group",
                  "value": {
                    "studentId": 501,
                    "classGroupId": 10,
                    "schoolYearId": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Enrollments"
        ]
      }
    },
    "/enrollments/{id}": {
      "get": {
        "operationId": "EnrollmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Enrollments"
        ]
      },
      "delete": {
        "operationId": "EnrollmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Enrollments"
        ]
      }
    },
    "/enrollments/{id}/deactivate": {
      "patch": {
        "operationId": "EnrollmentsController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Enrollments"
        ]
      }
    },
...
    "/grades/{id}": {
      "get": {
        "operationId": "GradesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Grades"
        ]
      },
      "patch": {
        "operationId": "GradesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGradeDto"
              },
              "examples": {
                "default": {
                  "summary": "Update grade comment",
                  "value": {
                    "comment": "Improved after retest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role teacher, admin"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Grades"
        ]
      },
      "delete": {
        "operationId": "GradesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role teacher, admin"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Grades"
        ]
      }
    },
    "/grade-schemes": {
      "get": {
        "operationId": "GradeSchemesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "grade-schemes"
        ]
      },
      "post": {
        "operationId": "GradeSchemesController_create",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "grade-schemes"
        ]
      }
    },
    "/grade-schemes/{id}": {
      "get": {
        "operationId": "GradeSchemesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "grade-schemes"
        ]
      },
      "patch": {
        "operationId": "GradeSchemesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "grade-schemes"
        ]
      },
      "delete": {
        "operationId": "GradeSchemesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "grade-schemes"
        ]
      }
    },
    "/grade-scheme-values": {
      "get": {
        "operationId": "GradeSchemeValuesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "grade-scheme-values"
        ]
      },
      "post": {
        "operationId": "GradeSchemeValuesController_create",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "grade-scheme-values"
        ]
      }
    },
    "/grade-scheme-values/{id}": {
      "get": {
        "operationId": "GradeSchemeValuesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "grade-scheme-values"
        ]
      },
      "patch": {
        "operationId": "GradeSchemeValuesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "grade-scheme-values"
        ]
      },
      "delete": {
        "operationId": "GradeSchemeValuesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "grade-scheme-values"
        ]
      }
    },
...
  },
  "info": {
    "title": "School Man API",
    "description": "Generated OpenAPI specification for School Man backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEnrollmentDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "number",
            "description": "Identifier of the student being enrolled",
            "example": 501,
            "minimum": 1
          },
          "classGroupId": {
            "type": "number",
            "description": "Identifier of the class group the student joins",
            "example": 32,
            "minimum": 1
          },
          "schoolYearId": {
            "type": "number",
            "description": "Identifier of the school year for the enrollment",
            "example": 2026,
            "minimum": 1
          }
        },
        "required": [
          "studentId",
          "classGroupId",
          "schoolYearId"
        ]
      },
...
      "UpdateGradeDto": {
        "type": "object",
        "properties": {
          "mark": {
            "type": "string",
            "description": "Updated letter mark for the student",
            "example": "B",
            "enum": [
              "S",
              "A",
              "B",
              "J"
            ]
          },
          "comment": {
            "type": "string",
            "description": "Updated comment for the grade",
            "example": "Improved after remedial sessions.",
            "maxLength": 500
          }
        }
      },
...
    }
  }
}